Refactor client for CRDs.
