+++ src/KubeClient/ClientFactoryExtensions.cs
@@ -198,24 +198,5 @@ namespace KubeClient
                 client => new DeploymentClientV1Beta1(client)
             );
         }
-
-        /// <summary>
-        ///     Get the Kubernetes CustomResourceDefinitions (v1beta1) resource client.
-        /// </summary>
-        /// <param name="kubeClient">
-        ///     The Kubernetes API client.
-        /// </param>
-        /// <returns>
-        ///     The resource client.
-        /// </returns>
-        public static CustomResourceDefinitionClientV1Beta1 CustomResourceDefinitionsV1Beta1(this KubeApiClient kubeClient)
-        {
-            if (kubeClient == null)
-                throw new ArgumentNullException(nameof(kubeClient));
-            
-            return kubeClient.ResourceClient(
-                client => new CustomResourceDefinitionClientV1Beta1(client)
-            );
-        }
     }
 }
